Clean up parameters and constructors. (#28819)
